+++ pkg/aws/awscluster/validate_awscluster.go
@@ -1,6 +1,7 @@
 package awscluster
 
 import (
+	"context"
 	"fmt"
 
 	"github.com/giantswarm/apiextensions/v3/pkg/annotation"
@@ -42,6 +43,11 @@ func (v *Validator) Validate(request *admissionv1.AdmissionRequest) (bool, error
 
 	if _, _, err := validator.Deserializer.Decode(request.Object.Raw, nil, &awsCluster); err != nil {
 		return false, microerror.Maskf(parsingFailedError, "unable to parse awscluster: %v", err)
+	}
+
+	err = aws.ValidateOrganizationLabelContainsExistingOrganization(context.Background(), v.k8sClient.CtrlClient(), &awsCluster)
+	if err != nil {
+		return false, microerror.Mask(err)
 	}
 
 	err = v.AWSClusterAnnotationMaxBatchSizeIsValid(awsCluster)