+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Validate existing organization label when cluster is created.
+
 ## [2.10.0] - 2021-04-06
 
 ### Added